+++ CleverPush.iOS.Binding/CleverPush.linkwith.cs
@@ -1,3 +1,3 @@
 using ObjCRuntime;
 
-[assembly: LinkWith ("CleverPush.a", ForceLoad = true, Frameworks= "SystemConfiguration UserNotifications WebKit CoreGraphics UIKit")]
+[assembly: LinkWith("CleverPush.a", ForceLoad = true, Frameworks = "CoreGraphics, QuartzCore, CoreFoundation, UIKit, SystemConfiguration, UIKit, UserNotifications, StoreKit, WebKit, JavascriptCore, SafariServices")]