+++ CleverPush.iOS.Binding/Properties/AssemblyInfo.cs
@@ -26,12 +26,10 @@ using ObjCRuntime;
 // The form "{Major}.{Minor}.*" will automatically update the build and revision,
 // and "{Major}.{Minor}.{Build}.*" will update just the revision.
 
-[assembly: AssemblyVersion ("1.2.3")]
+[assembly: AssemblyVersion ("1.3.0")]
 
 // The following attributes are used to specify the signing key for the assembly,
 // if desired. See the Mono documentation for more information about signing.
 
 //[assembly: AssemblyDelaySign(false)]
 //[assembly: AssemblyKeyFile("")]
-
-[assembly: LinkWith("CleverPush.a", LinkTarget.Simulator, Frameworks = "CoreGraphics, QuartzCore, CoreFoundation, UIKit, SystemConfiguration, UIKit, UserNotifications, StoreKit, WebKit, JavascriptCore, SafariServices")]